requirements.txt
fastapi
uvicorn[standard]
httpx[http2]
redis
pydantic
pydantic-settings
apscheduler
pytest
pytest-asyncio

SQLAlchemy
python-dateutil
asyncpg
greenlet
pandas
openpyxl
